Implement user registration and login functionality, add user profile routes, and integrate file upload middleware
Signed by Tanuj Keshattiwar
