+++ backend/package.json
@@ -16,6 +16,8 @@
     "dotenv": "^17.3.1",
     "express": "^5.2.1",
     "jsonwebtoken": "^9.0.3",
-    "mongoose": "^9.2.4"
+    "mongoose": "^9.2.4",
+    "multer": "^2.1.1",
+    "uuid": "^13.0.0"
   }
 }
