+++ backend/src/app.js
@@ -1,6 +1,9 @@
 import express from 'express';
 import userLoginRoute from './routes/auth.route.js';
 import cookieParser from 'cookie-parser';
+import userRoutes from './routes/user.routes.js'
+
+
 const app = express();
 
 app.get('/', (req, res) => {
@@ -12,5 +15,6 @@ app.use(express.urlencoded({ extended: true }));
 app.use(cookieParser());
 
 app.use("/api/auth", userLoginRoute);
+app.use("/api/users",userRoutes);
 
 export default app;
